feat: added react router dom and rearanged components to meet router specifications
