+++ src/App.css
@@ -1,4 +1,4 @@
-.App {
+.wrapper {
   display: flex;
   flex-direction: column;
   justify-content: center;
@@ -18,7 +18,7 @@ button {
   height: 40px;
 }
 
-.App button {
+button {
   margin: 20px;
   width: 200px;
 }
