+++ src/Components/Layout/Layout.tsx
@@ -1,0 +1,11 @@
+import { Outlet } from 'react-router-dom';
+import { Header } from '../Header/Header';
+
+export const Layout = () => {
+  return (
+    <div className="layout">
+      <Header />
+      <Outlet />
+    </div>
+  );
+};
